+++ .changeset/cool-cooks-act.md
@@ -1,0 +1,6 @@
+---
+'myst-directives': patch
+'myst-transforms': patch
+---
+
+Move includeDirective transform to myst-transforms and make it generic for use in JupyterLab
