+++ packages/myst-transforms/src/index.ts
@@ -40,6 +40,7 @@ export {
 export { joinGatesPlugin, joinGatesTransform } from './joinGates.js';
 export { glossaryPlugin, glossaryTransform } from './glossary.js';
 export { abbreviationPlugin, abbreviationTransform } from './abbreviations.js';
+export { includeDirectivePlugin, includeDirectiveTransform } from './include.js';
 
 // Enumeration
 export type { IReferenceState, NumberingOptions, ReferenceKind } from './enumerate.js';
